% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthor.r
\docType{data}
\name{coauthors}
\alias{coauthors}
\title{Example data of author list}
\description{
Example data of author list
}
\references{
Keita Fukasawa, Takahiro Morosawa, Yoshihiro Nakashima, Shun Takagi, Takumasa Yokoyama, Masaki Ando, Hayato Iijima, Masayuki Saito, Nao Kumada, Kahoko Tochigi, Akira Yoshioka, Satsuki Funatsu, Shinsuke Koike, Hiroyuki Uno, Takaaki Enomoto, William McShea, Roland Kays (2025) Snapshot Japan 2023: the first camera trap dataset under a globally standardised protocol in Japan. Biodiversity Data Journal in press.
}
\author{
Keita Fukasawa \email{k.fukasawa37@gmail.com}
}
\keyword{data}
